refactor: move error modules to core
